+++ test/unit/common_clj/test/helper/core_test.clj
@@ -1,0 +1,9 @@
+(ns common-clj.test.helper.core-test
+  (:require [clojure.test :refer :all]
+            [common-clj.test.helper.core :as test.helper]
+            [matcher-combinators.test :refer [match?]]))
+
+(deftest uuid-test
+  (testing "that we can generate random Uuids"
+    (is (match? uuid?
+                (test.helper/uuid)))))