+++ test/unit/common_clj/test/helper_test.clj
@@ -1,9 +1,0 @@
-(ns common-clj.test.helper-test
-  (:require [clojure.test :refer :all]
-            [common-clj.test.helper :as test.helper]
-            [matcher-combinators.test]))
-
-(deftest uuid-test
-  (testing "that we can generate random Uuids"
-    (is (match? uuid?
-           (test.helper/uuid)))))